+++ cuda_bindings/pyproject.toml
@@ -8,7 +8,8 @@ requires = [
     "cython>=3.2,<3.3", 
     "pyclibrary>=0.1.7"
 ]
-build-backend = "setuptools.build_meta"
+build-backend = "build_hooks"
+backend-path = ["."]
 
 [project]
 name = "cuda-bindings"
